fix(): support live reload in start mode
